Started the base of the solver
Added the constructor that takes in a int?[,] so the solve cell can be null.
I also made the solve method and methods to test by addition, subtraction, division and multiplication. They are unfinished but the base for going through rows is there. Just need to go through and compare the current number and the next number in a row and then check columns